### 2019-04-23 #1401 @katyhuff
I've found some great folks. Now @prmiles , let's ping a few fitting reviewers, with python, statistics (ideally bayesian), and some computational physics expertise. We can start the review once two or three of them accept the request to review.
