### 2019-04-23 #1401 @katyhuff
Dear @ivergara :

I have identified you as a fitting reviewer for this paper, which will require experts with python, statistics (ideally bayesian), and some computational physics expertise. Are you able to and interested in taking on this JOSS review?

Title: pymcmcstat: A Python Package for Bayesian Inference Using Delayed Rejection Adaptive Metropolis
Summary:  A Python program for running Markov Chain Monte Carlo (MCMC) simulations.  Included in this package is the ability to use different Metropolis based sampling techniques including Metropolis-Hastings, Adaptive-Metropolis, Delayed-Rejection, and Delayed Rejection Adaptive Metropolis. This Python package is an adaptation of the MATLAB toolbox mcmcstat.
Article Proof: https://github.com/openjournals/joss-papers/blob/joss.01401/joss.01401/10.21105.joss.01401.pdf
Submitting author: @prmiles (Paul Miles)
Repository: https://github.com/prmiles/pymcmcstat
Version: v1.7.0
