### 2019-04-23 #1401 @djmitche
I have good experience in Python and minimal statistical experience.  On that basis, I'm happy to help.
