### 2019-04-24 #1401 @ivergara
@katyhuff unfortunately I cannot take this review, Bayesian statistics falls (at the moment) too far from my knowledge corpus to be able to provide a good feedback on how the package works.
