### 2019-04-26 bcf0c8f
updated version number for dev branch
