### 2019-04-27 4a3d0dd
updated docs to provide more details regarding acceptance ratio
- outlined likelihood and prior functions
