### 2019-04-27 6320aba
fixed flake8 requirements
