### 2019-04-29 #1401 @ejolly
@katyhuff This package looks really great and I would be happy to help review however this week and for the next 3 weeks I'll be traveling so I likely won't be able to assist reviewing for about a month. If that timeline works I'm happy to conduct the review then. Otherwise, I'd suggest finding someone else who can review on a shorter timeline. Apologies!
