### 2019-04-29 8d66ee2
added full default gaussian priorfunction
- evaluates np.exp(-1./2*pf)
