### 2019-04-29 d7b5a66
added acceptance ratio function that uses general likelihoods and priors
