### 2019-04-29 d34bfa1
added LikelihoodFunction class file
