### 2019-04-29 9c67c25
update likelihood evaluation
