### 2019-04-29 #1417 @whedon
[REVIEW]: pymcmcstat: A Python Package for Bayesian Inference Using Delayed Rejection Adaptive Metropolis
**Submitting author:** @prmiles (<a href="http://orcid.org/0000-0002-7501-5114">Paul Miles</a>)
**Repository:** <a href="https://github.com/prmiles/pymcmcstat" target ="_blank">https://github.com/prmiles/pymcmcstat</a>
**Version:** v1.7.1
