### 2019-04-30 7eef395
added posterior and log posterior ratio acceptance tests
- added corresponding unit tests
