+++ test/samplers/test_utilities.py
@@ -9,6 +9,7 @@ from pymcmcstat.samplers.utilities import sample_candidate_from_gaussian_proposa
 from pymcmcstat.samplers.utilities import is_sample_outside_bounds
 from pymcmcstat.samplers.utilities import acceptance_test
 from pymcmcstat.samplers.utilities import set_outside_bounds
+from pymcmcstat.samplers.utilities import log_posterior_ratio_acceptance_test as lprat
 from pymcmcstat.structures.ParameterSet import ParameterSet
 import unittest
 from mock import patch
@@ -39,19 +40,39 @@ class OutsideBounds(unittest.TestCase):
         self.assertFalse(is_sample_outside_bounds(theta = np.array([0.5, 2]), lower_limits = np.array([0,1]), upper_limits=np.array([1,2])), msg = 'p2 on upp lim')
         self.assertFalse(is_sample_outside_bounds(theta = np.array([0.5, 1.5]), lower_limits = np.array([0,1]), upper_limits=np.array([1,2])), msg = 'all outside')
 
+
 # --------------------------
 class Acceptance(unittest.TestCase):
     
     def test_accept_alpha_gt_1(self):
-        self.assertEqual(acceptance_test(alpha = 1.1), 1, msg = 'accept alpha >= 1')
-        self.assertEqual(acceptance_test(alpha = 1), 1, msg = 'accept alpha >= 1')
-        self.assertEqual(acceptance_test(alpha = -0.1), 0, msg = 'Reject alpha <= 0')
-        self.assertEqual(acceptance_test(alpha = 0), 0, msg = 'Reject alpha <= 0')
+        self.assertEqual(acceptance_test(alpha=1.1, logpost=False), 1,
+                         msg='accept alpha >= 1')
+        self.assertEqual(acceptance_test(alpha=1, logpost=False), 1,
+                         msg='accept alpha >= 1')
+        self.assertEqual(acceptance_test(alpha=-0.1, logpost=False), 0,
+                         msg='Reject alpha <= 0')
+        self.assertEqual(acceptance_test(alpha=0, logpost=False), 0,
+                         msg='Reject alpha <= 0')
         np.random.seed(0)
-        self.assertEqual(acceptance_test(alpha = 0.4), 0, msg = 'Reject alpha < u (0.4 < 0.5488135)')
+        self.assertEqual(acceptance_test(alpha=0.4, logpost=False), 0,
+                         msg='Reject alpha < u (0.4 < 0.5488135)')
         np.random.seed(0)
-        self.assertEqual(acceptance_test(alpha = 0.6), 1, msg = 'Accept alpha > u (0.6 > 0.5488135)')
-        
+        self.assertEqual(acceptance_test(alpha=0.6, logpost=False), 1,
+                         msg='Accept alpha > u (0.6 > 0.5488135)')
+
+    def test_log_post_alpha(self):
+        self.assertEqual(acceptance_test(alpha=0.1, logpost=True), 1,
+                         msg='accept alpha >= 0')
+        self.assertEqual(acceptance_test(alpha=0, logpost=True), 1,
+                         msg='accept alpha >= 0')
+        np.random.seed(0)
+        self.assertEqual(acceptance_test(alpha=np.log(0.4), logpost=True), 0,
+                         msg='Reject alpha < log(u) (-0.916 < -0.600)')
+        np.random.seed(0)
+        self.assertEqual(acceptance_test(alpha=np.log(0.6), logpost=True), 1,
+                         msg='Accept alpha > log(u) (-0.511 > -0.600)')
+
+
 # --------------------------
 class SetOutsideBounds(unittest.TestCase):
     def test_set_outsidebounds(self):
[…]
