### 2019-04-30 5084a39
updated documentation and added markers where likelihood features need to be added
