### 2019-04-30 8638c44
import likelihood function class
