### 2019-04-30 55a5d87
initialized likelihood function object
