### 2019-04-30 d28d07e
Update CHANGELOG.rst
