### 2019-04-30 6910c50
condensed parameter structure setup in initialization
