### 2019-04-30 ed7b57f
cleaned up model/options dependencies
