### 2019-04-30 571cdb8
cleaned up model/option dependencies
