+++ test/settings/test_SimulationOptions.py
@@ -24,12 +24,12 @@ class DisplaySimulationOptions(unittest.TestCase):
     def test_print_these_none(self):
         SO = setup_options()
         print_these = SO.display_simulation_options(print_these = None)
-        self.assertEqual(print_these, ['nsimu', 'adaptint', 'ntry', 'method', 'printint', 'lastadapt', 'drscale', 'qcov'], msg = 'Default print keys')
+        self.assertEqual(print_these, ['nsimu', 'adaptint', 'ntry', 'method', 'printint', 'lastadapt', 'drscale', 'qcov'], msg='Default print keys')
         
     def test_print_these_not_none(self):
         SO = setup_options()
         print_these = SO.display_simulation_options(print_these = ['nsimu'])
-        self.assertEqual(print_these, ['nsimu'], msg = 'Specified print keys')
+        self.assertEqual(print_these, ['nsimu'], msg='Specified print keys')
         
 # --------------------------
 class CheckDependentOptions(unittest.TestCase):
@@ -37,29 +37,38 @@ class CheckDependentOptions(unittest.TestCase):
     def test_dodram(self):
         SO = setup_options(ntry = 1)
         SO._check_dependent_simulation_options(model = MS)
-        self.assertEqual(SO.dodram, 0, msg = 'DRAM turned off because ntry <= 1')
+        self.assertEqual(SO.dodram, 0, msg='DRAM turned off because ntry <= 1')
         SO = setup_options(ntry = 2)
         SO._check_dependent_simulation_options(model = MS)
-        self.assertEqual(SO.dodram, 1, msg = 'DRAM turned on because ntry > 1')
+        self.assertEqual(SO.dodram, 1, msg='DRAM turned on because ntry > 1')
         
     def test_lastadapt(self):
         SO = setup_options(lastadapt = 0)
-        SO._check_dependent_simulation_options(model = MS)
-        self.assertEqual(SO.lastadapt, SO.nsimu, msg = 'lastadapt set to nsimu')
-        SO.define_simulation_options(lastadapt = 10)
-        SO._check_dependent_simulation_options(model = MS)
-        self.assertEqual(SO.lastadapt, 10, msg = 'lastadapt unchanged')
+        SO._check_dependent_simulation_options(model=MS)
+        self.assertEqual(SO.lastadapt, SO.nsimu,
+                         msg='lastadapt set to nsimu')
+        SO.define_simulation_options(lastadapt=10)
+        SO._check_dependent_simulation_options(model=MS)
+        self.assertEqual(SO.lastadapt, 10,
+                         msg='lastadapt unchanged')
         
     def test_printint(self):
         SO = setup_options(printint = None)
-        SO._check_dependent_simulation_options(model = MS)
-        self.assertEqual(SO.printint, max(100,min(1000,SO.adaptint)), msg = 'printint was updated')
+        SO._check_dependent_simulation_options(model=MS)
+        self.assertEqual(SO.printint, max(100,min(1000,SO.adaptint)),
+                         msg='printint was updated')
         
     def test_updatesigma(self):
-        SO = setup_options(updatesigma = False)
+        SO = setup_options(updatesigma=False)
         MS.N0 = [1.]
-        SO._check_dependent_simulation_options(model = MS)
-        self.assertTrue(SO.updatesigma, msg = 'updatesigma turned on because N0 not empty')
+        SO._check_dependent_simulation_options(flag=MS.N0 is not None)
+        self.assertTrue(SO.updatesigma,
+                        msg='updatesigma turned on because N0 not empty')
+        SO = setup_options(updatesigma=False)
+        MS.N0 = None
+        SO._check_dependent_simulation_options(flag=MS.N0 is not None)
[…]
