### 2019-04-30 94d3737
updated MCMC to comply with new likelihood call statements
