### 2019-04-30 2611c6a
updated likelihood class to have its own sum-of-squares default attributes
