### 2019-04-30 18476b5
added like_object to input
- now calculate likelihood
- need to phase out ssq calcs
