### 2019-04-30 e88a38c
added like, loglike, and logprior attributes
