+++ pymcmcstat/structures/ParameterSet.py
@@ -20,9 +20,14 @@ class ParameterSet:
         * **sigma2** (:class:`~numpy.ndarray`): Observation errors.
         * **alpha** (:py:class:`float`): Result from evaluating likelihood function.
     '''
-    def __init__(self, theta=None, ss=None, prior=None, sigma2=None, alpha=None):
+    def __init__(self, theta=None, ss=None,
+                 sigma2=None, alpha=None, like=None,
+                 prior=None, logprior=None, loglike=None):
         self.theta = theta
         self.ss = ss
-        self.prior = prior
         self.sigma2 = sigma2
         self.alpha = alpha
+        self.like = like
+        self.loglike = loglike
+        self.prior = prior
+        self.logprior = logprior
