### 2019-04-30 8890fbb
added tests for Likelihood function class
