### 2019-04-30 bbb3801
updated samplers
