### 2019-04-30 bcca671
updated unpack_set test function
