### 2019-04-30 d652aeb
updated error variance estimator to utilize custom samplers method
