+++ test/procedures/test_ErrorVarianceEstimator.py
@@ -16,18 +16,18 @@ import numpy as np
 class InitializeEVE(unittest.TestCase):
 
     def test_init_EVE(self):
-        EVE = ErrorVarianceEstimator()
+        EVE = ErrorVarianceEstimator(model=None, nsimu=None)
         self.assertTrue(hasattr(EVE, 'description'))
         
 # --------------------------
 class UpdateEVE(unittest.TestCase):
     
     def test_eve_update_with_nsos_1(self):
-        model, __, parameters, data = gf.setup_mcmc()
+        model, options, parameters, data = gf.setup_mcmc()
         SOS = SumOfSquares(model = model, data = data, parameters = parameters)
         theta = np.array([2., 5.])
         ss = SOS.evaluate_sos_function(theta = theta)
-        EVE = ErrorVarianceEstimator()
+        EVE = ErrorVarianceEstimator(model, options.nsimu)
         sigma2 = EVE.update_error_variance(sos = ss, model = model)
         self.assertTrue(isinstance(sigma2, np.ndarray), msg = 'Expect array return.')
         self.assertEqual(sigma2.size, 1, msg = 'Size of array is 1')
@@ -38,7 +38,7 @@ class UpdateEVE(unittest.TestCase):
         nsos = 2
         model._check_dependent_model_settings_wrt_nsos(nsos = nsos)
         ss = np.array([0.1, 0.2])
-        EVE = ErrorVarianceEstimator()
+        EVE = ErrorVarianceEstimator(model, options.nsimu)
         sigma2 = EVE.update_error_variance(sos = ss, model = model)
         self.assertTrue(isinstance(sigma2, np.ndarray), msg = 'Expect array return.')
         self.assertEqual(sigma2.size, 2, msg = 'Size of array is 2')
@@ -48,14 +48,14 @@ class UpdateEVE(unittest.TestCase):
 class Gammar(unittest.TestCase):
     
     def test_gammar_for_a_0(self):
-        EVE = ErrorVarianceEstimator()
+        EVE = ErrorVarianceEstimator(model=None, nsimu=None)
         m = 4
         n = 3
         ret = EVE.gammar(m = m, n = 3, a = 0)
         self.assertTrue(np.array_equal(ret, np.zeros([m,n])), msg = 'Expect equal array return.')
         
     def test_gammar_for_a_lt_0(self):
-        EVE = ErrorVarianceEstimator()
+        EVE = ErrorVarianceEstimator(model=None, nsimu=None)
         m = 4
         n = 3
         ret = EVE.gammar(m = m, n = 3, a = -10.)
@@ -65,7 +65,7 @@ class Gammar(unittest.TestCase):
 class Gammar_MT1(unittest.TestCase):
     
     def test_gammar_for_a_lt_1(self):
-        EVE = ErrorVarianceEstimator()
+        EVE = ErrorVarianceEstimator(model=None, nsimu=None)
         y = EVE._gammar_mt1(a = 0.1, b = 1)
         self.assertTrue(isinstance(y, np.ndarray), msg = 'Expect equal array return.')
         
