### 2019-05-01 afddf44
updated error variance estimator to be compatible with vector sum-of-squares
- still need to test features when turned off
