+++ pymcmcstat/MCMC.py
@@ -114,7 +114,8 @@ class MCMC:
         self.__generate_simulation_results()
         if self.simulation_options.save_to_json is True:
             if self.simulation_results.basic is True:  # check that results structure has been created
-                self.simulation_results.export_simulation_results_to_json_file(results=self.simulation_results.results)
+                self.simulation_results.export_simulation_results_to_json_file(
+                        results=self.simulation_results.results)
         self.mcmcplot = MCMCPlotting.Plot()
         self.PI = PredictionIntervals()
         self.chainstats = ChainStatistics.chainstats
@@ -151,7 +152,6 @@ class MCMC:
                 res = RS.load_json_object(self.simulation_options.json_restart_file)
                 self.parameters._results_to_params(res, 1)
                 self.simulation_options.qcov = np.array(res['qcov'])
-
             self.__chain_index = 0  # start index at zero
             self._initialize_simulation()
             self.__initialize_chains(
@@ -193,23 +193,16 @@ class MCMC:
                 mu=self.parameters._thetamu[self.parameters._parind[:]],
                 sigma=self.parameters._thetasigma[self.parameters._parind[:]])
         # ---------------------
-        # Check if user defined likelihood
+        # Check if updatesigma
         if self.simulation_options.updatesigma is True:
             self.custom_samplers.append(
                     ErrorVarianceEstimator(
                             self.model_settings,
-                            self.simulation_options.nsimu))
-        # ---------------------
-        # Setup custom samplers
-        self.custom_sampler_output = []
-        if len(self.custom_samplers) > 0:
-            for ii, cs in enumerate(self.custom_samplers):
-                self.custom_sampler_output.append(cs.setup())
+                            self.simulation_options.nsimu,
+                            self.simulation_options.s2chainfile))
         # ---------------------
         # Define initial parameter set
         self.__setup_initial_parameter_set()
-        # recheck certain values in model settings that are dependent on the output of the sos function
-        self.model_settings._check_dependent_model_settings_wrt_nsos(nsos=len(self.__initial_set.ss))
         # ---------------------
         # Update variables covariance adaptation
         self._covariance._update_covariance_settings(self.__initial_set.theta)
@@ -223,7 +216,7 @@ class MCMC:
 
         Args:
             * **chainind** (:py:class:`int`): Where to store initial parameter value
-            * **nsimu** (:py:class:`int`): Number of parameter samples to simulate.  Default is 1e4.
+            * **nsimu** (:py:class:`int`): Number of parameter samples to simulate.
             * **npar** (:py:class:`int`): Number of parameters being sampled.
             * **nsos** (:py:class:`int`): Length of output from sum-of-squares function
             * **updatesigma** (:py:class:`bool`): Flag for updating measurement error variance. \
@@ -236,15 +229,12 @@ class MCMC:
         # Save initialized values to chain, s2chain, sschain
         self.__chain[chainind, :] = self.__initial_set.theta
         self.__sschain[chainind, :] = self.__initial_set.ss
-
         self.__chains = []
         self.__chains.append(dict(file=self.simulation_options.chainfile, mtx=self.__chain))
         self.__chains.append(dict(file=self.simulation_options.sschainfile, mtx=self.__sschain))
-
         if updatesigma:
             self.__s2chain = np.zeros([nsimu, nsos])
             self.__s2chain[chainind, :] = sigma2
-            self.__chains.append(dict(file=self.simulation_options.s2chainfile, mtx=self.__s2chain))
         else:
             self.__s2chain = None
 
@@ -284,10 +274,11 @@ class MCMC:
         savecount = 0  # save counter
         lastbin = 0  # initialize bin counter
[…]
