+++ pymcmcstat/procedures/ErrorVarianceEstimator.py
@@ -18,10 +18,12 @@ class ErrorVarianceEstimator:
         * :meth:`~gammar`
         * :meth:`~gammar_mt`
     '''
-    def __init__(self, model, nsimu):
+    def __init__(self, model, nsimu, name='s2chainfile'):
         self.description = 'Estimate error variance.'
         self.model = model
         self.nsimu = nsimu
+        self.name = name
+        self.save_chain = True
 
     def setup(self):
         self.status = 'setup'
@@ -29,14 +31,14 @@ class ErrorVarianceEstimator:
         self.counter = 0
         self.sigma2[self.counter, :] = self.model.sigma2
         self.chains = []
-        self.chains.append(dict(file='sigma2chain', mtx=self.sigma2))
+        self.chains.append(dict(file=self.name, mtx=self.sigma2))
         return self.model.sigma2
 
     def update(self, current_set, **kwargs):
         self.status = 'updating'
         self.counter += 1
         sigma2 = self.update_error_variance(current_set.ss, self.model)
-        self.sigma2[self.counter, 0] = sigma2
+        self.sigma2[self.counter, :] = sigma2.reshape(sigma2.size, )
         return sigma2
 
     def update_error_variance(self, sos, model):
