### 2019-05-02 prmiles/pymcmcstat#41 @djmitche
Prints version on import
**Describe the bug**
pymcmcstat prints its version when imported

**Expected behavior**
Imports occur without side-effects, and in particular no output.

It's pretty atypical for a Python module to write to stdout on import, and would make it a bit awkward to fit this into a larger command-line application, for example (since it would be included in that application's output).

Refs https://github.com/openjournals/joss-reviews/issues/1417
