### 2019-05-02 prmiles/pymcmcstat#42 @djmitche
Add CONTRIBUTING.md
While the README and documentation provide good guidance for getting support and reporting bugs, there is little information on how to contribute to the project.  How should proposed changes be submitted?  What can contributors expect in terms of review?  What guidelines must they follow in modifying the library?  How do they run tests?  Etc..

Refs https://github.com/openjournals/joss-reviews/issues/1417
