### 2019-05-02 #1417 @djmitche
I'm happy aside from the above-referenced two issues.

While this does not have a very "pythonic" interface, that's justified in the documentation as an effort to match the MATLAB functionality, and I think will match the expectations of users of this sort of application nicely.  In particular, I'd highlight that many methods "display" data by printing to stdout rather than return it, and that running the simulation modifies the MCMC object itself, from which results must be fetched, rather than leaving the object unchanged and returning the simulation results.    It seems that the displayed results are also available programmatically if necessary (e.g., ModelSettings), so this would not be an impediment to using the library as a component of a larger application.
