### 2019-05-02 #1417 @katyhuff
Thank you @djmitche for your prompt review! You've made some very good points that perhaps @prmiles could consider as improvements to the package!
