### 2019-05-02 c4c7df4
updated init file to suppress printing version number
