+++ pymcmcstat/__version__.py
@@ -1,1 +1,1 @@
-__version__ = "1.7.0"
+__version__ = "1.7.1rc1"
