+++ pymcmcstat/__init__.py
@@ -1,2 +1,1 @@
 from .__version__ import __version__
-print('pymcmcstat version: {}'.format(__version__))
