### 2019-05-02 8c07b81
moved version attribute to init file
