### 2019-05-02 324e2b7
updated how setup.py accesses version attribute
- reads in from init file using re
