+++ setup.py
@@ -1,5 +1,7 @@
 from setuptools import setup, find_packages
 import codecs
+import os
+import re
 
 def read(fname):
     with codecs.open(fname, 'r', 'latin') as f:
@@ -11,15 +13,20 @@ this_directory = path.abspath(path.dirname(__file__))
 with codecs.open(path.join(this_directory, 'README.rst'), encoding='utf-8') as f:
     long_description = f.read()
     
-# read in version number
-version_dummy = {}
-exec(read('pymcmcstat/__version__.py'), version_dummy)
-__version__ = version_dummy['__version__']
-del version_dummy
+
+def get_version():
+    VERSIONFILE = os.path.join('pymcmcstat', '__init__.py')
+    lines = open(VERSIONFILE, 'rt').readlines()
+    vgx = '^__version__ = \"\d+\.\d+\.\d.*\"'
+    for line in lines:
+        mo = re.search(vgx, line, re.M)
+        if mo:
+            return mo.group().split('"')[1]
+    raise RuntimeError('Unable to find version in %s.' % (VERSIONFILE,))
 
 setup(
     name='pymcmcstat',
-    version=__version__,
+    version=get_version(),
     description=('A library to perform MCMC simulations using DRAM'),
     long_description=long_description,
     url='https://github.com/prmiles/pymcmcstat',
