### 2019-05-02 9fe19a1
added context manager for reading in init file
