### 2019-05-02 1f93d10
removed version file
