### 2019-05-02 prmiles/pymcmcstat#41 @djmitche
Thanks for the tests too :)
