### 2019-05-02 prmiles/pymcmcstat#41 @prmiles
Thank you for pointing out the issue.  I am still somewhat new to Python :)
