### 2019-05-03 cfc1688
Create CONTRIBUTING.md
