### 2019-05-03 174e63b
Update CONTRIBUTING.md
