### 2019-05-03 4432748
Update and rename CONTRIBUTING.md to CONTRIBUTING.rst
