+++ CONTRIBUTING.rst
@@ -1,4 +1,5 @@
-# Guidelines for Contributing
+Guidelines for Contributing
+===========================
 
 Individual or groups are very welcome to contribute to pymcmcstat.  There are four main ways of contributing to the pymcmcstat project:
 
@@ -7,67 +8,71 @@ Individual or groups are very welcome to contribute to pymcmcstat.  There are fo
 3. Contributing or improving the documentation (`doc`).
 4. Submitting issues related to bugs or desired enhancements.
 
-# Opening issues
+Opening issues
+==============
 
-The best way to announce issues is via [Github Issue Tracker](https://github.com/prmiles/pymcmcstat/issues).  Please verify that your issue is not being currently addressed by other issues or pull requests by using the GitHub search tool to look for key words in the project issue tracker.
+The best way to announce issues is via `Github Issue Tracker <https://github.com/prmiles/pymcmcstat/issues>`_.  Please verify that your issue is not being currently addressed by other issues or pull requests by using the GitHub search tool to look for key words in the project issue tracker.
 
-# Contributing code via pull requests
+Contributing code via pull requests
+===================================
 
 If you would like to address an issue directly, users are strongly encouraged to submit patches for new or existing issues via pull requests.  This is especially appreciated for simple fixes like typos or tweaks to documentation.
 
-Contributors are also encouraged to contribute new code to enhance pymcmcstat's functionality, also via pull requests. Please consult the [pymcmcstat documentation](https://pymcmcstat.readthedocs.io/) to ensure that any new contribution does not strongly overlap with existing functionality.
+Contributors are also encouraged to contribute new code to enhance pymcmcstat's functionality, also via pull requests. Please consult the `pymcmcstat documentation <https://pymcmcstat.readthedocs.io/>`_ to ensure that any new contribution does not strongly overlap with existing functionality.
 
-The recommended workflow for contributing to pymcmcstat is to fork the [GitHub repository](https://github.com/prmiles/pymcmcstat), clone it to your local machine, and develop on a feature branch.
+The recommended workflow for contributing to pymcmcstat is to fork the `GitHub repository <https://github.com/prmiles/pymcmcstat>`_, clone it to your local machine, and develop on a feature branch.
 
-## Steps:
+Steps:
+------
 
-1. Fork the [project repository](https://github.com/prmiles/pymcmcstat) by clicking on the 'Fork' button near the top right of the main repository page. This creates a copy of the code under your GitHub user account.
+1. Fork the `project repository <https://github.com/prmiles/pymcmcstat>`_ by clicking on the 'Fork' button near the top right of the main repository page. This creates a copy of the code under your GitHub user account.
 
 2. Clone your fork of the pymcmcstat repo from your GitHub account to your local disk, and add the base repository as a remote:
 
-   ```bash
+::
+
    $ git clone git@github.com:<your GitHub handle>/pymcmcstat.git
    $ cd pymcmcstat
    $ git remote add upstream git@github.com:prmiles/pymcmcstat.git
-   ```
 
 3. Create a ``feature`` branch to hold your development changes:
 
-   ```bash
+::
+
    $ git checkout -b my-feature
-   ```
 
-   Always use a ``feature`` branch. It's good practice to never routinely work on the ``master`` branch of any repository.
+Always use a ``feature`` branch. It's good practice to never routinely work on the ``master`` branch of any repository.
 
-4. Project requirements are in ``requirements.txt``. It is recommended that you use some type of [virtual environment](https://docs.python.org/3/tutorial/venv.html) for your code development.  To install the required packages run:
[…]
