### 2019-05-03 80fa519
Fix link to reference guide
