### 2019-05-03 #1417 @prmiles
Thank you @djmitche for your comments regarding the pymcmcstat package.  I have resolved the two issues you reported.
