### 2019-05-06 #1417 @djmitche
Indeed!  All boxes checked..
