### 2019-05-07 3e87174
updated how likelihood is passed between sampling objects
