### 2019-05-07 b2a3b09
updated procedures, prior and likelihood
- need to pass in custom arguments to prior function
