+++ pymcmcstat/procedures/PriorFunction.py
@@ -23,11 +23,9 @@ class PriorFunction:
 
         self.mu = mu
         self.sigma = sigma
-
-        # Setup prior function and evaluate
+        # Setup prior function
         if priorfun is None:
-            priorfun = self.default_priorfun
-
+            priorfun = self.default_gaussian_priorfun
         self.priorfun = priorfun  # function handle
 
     @classmethod
