### 2019-05-07 a2a42ba
updated posterior ratio calculator
