### 2019-05-07 0b4e37e
updated DR with respect to likelihood
