+++ pymcmcstat/samplers/utilities.py
@@ -130,6 +130,8 @@ def set_outside_bounds(next_set):
     next_set.alpha = 0
     next_set.prior = 0
     next_set.like = 0
+    next_set.logprior = -np.inf
+    next_set.loglike = -np.inf
     next_set.ss = np.inf
     outbound = True
 
@@ -169,17 +171,17 @@ def calculate_posterior_ratio(likestar, like,
     Returns:
         * **alpha** (:py:class:`float`): Acceptance ratio
     '''
-    print('likestar = {}'.format(likestar))
-    print('like = {}'.format(like))
-    print('priorstar = {}'.format(priorstar))
-    print('prior = {}'.format(prior))
+#    print('likestar = {}'.format(likestar))
+#    print('like = {}'.format(like))
+#    print('priorstar = {}'.format(priorstar))
+#    print('prior = {}'.format(prior))
     alpha = (likestar * priorstar)/(like * prior)
     return np.min([1, alpha])
 
 
 # -------------------------------------------
-def calculate_log_posterior_ratio(likestar, like,
-                               priorstar, prior):
+def calculate_log_posterior_ratio(loglikestar, loglike,
+                               logpriorstar, logprior):
     '''
     Calculate log posterior ratio:
 
@@ -210,9 +212,9 @@ def calculate_log_posterior_ratio(likestar, like,
     Returns:
         * **alpha** (:py:class:`float`): Acceptance ratio
     '''
-#    print('likestar = {}'.format(likestar))
-#    print('like = {}'.format(like))
-#    print('priorstar = {}'.format(priorstar))
-#    print('prior = {}'.format(prior))
-    alpha = (likestar * priorstar)/(like * prior)
-    return np.min([1, alpha])
+#    print('likestar = {}'.format(loglikestar))
+#    print('like = {}'.format(loglike))
+#    print('priorstar = {}'.format(logpriorstar))
+#    print('prior = {}'.format(logprior))
+    logalpha = loglikestar + logpriorstar - (loglike + logprior)
+    return np.min([0, logalpha])
