### 2019-05-08 fdb3b22
updated workflow to communicate with likelihood and prior objects
