+++ pymcmcstat/samplers/DelayedRejection.py
@@ -11,6 +11,7 @@ from ..structures.ParameterSet import ParameterSet
 from .utilities import sample_candidate_from_gaussian_proposal
 from .utilities import is_sample_outside_bounds, set_outside_bounds
 from .utilities import acceptance_test
+from .utilities import calculate_log_posterior_ratio
 
 
 class DelayedRejection:
@@ -24,7 +25,7 @@ class DelayedRejection:
     '''
     # -------------------------------------------
     def run_delayed_rejection(self, old_set, new_set, RDR, ntry, parameters,
-                              invR, sosobj, priorobj, likeobj, custom=None):
+                              invR, priorobj, likeobj, custom=None):
         '''
         Perform delayed rejection step - occurs in standard metropolis is not accepted.
 
@@ -35,7 +36,6 @@ class DelayedRejection:
             * **ntry** (:py:class:`int`): Number of DR steps to perform until rejection
             * **parameters** (:class:`~.ModelParameters`): Model parameters
             * **invR** (:class:`~numpy.ndarray`): Inverse Cholesky decomposition matrix
-            * **sosobj** (:class:`~.SumOfSquares`): Sum-of-Squares function
             * **priorobj** (:class:`~.PriorFunction`): Prior function
             * **likeobj** (:class:`~.LikelihoodFunction`): Likelihood function
 
@@ -52,11 +52,12 @@ class DelayedRejection:
             itry += 1  # update dr step index
             # initialize next step parameter set
             next_set = self.initialize_next_metropolis_step(
-                    npar=parameters.npar, old_theta=old_set.theta, sigma2=new_set.sigma2, RDR=RDR[itry-1])
+                    npar=parameters.npar, old_theta=old_set.theta, RDR=RDR[itry-1])
             # Reject points outside boundaries
-            outsidebounds = is_sample_outside_bounds(next_set.theta,
-                                                     parameters._lower_limits[parameters._parind[:]],
-                                                     parameters._upper_limits[parameters._parind[:]])
+            outsidebounds = is_sample_outside_bounds(
+                    next_set.theta,
+                    parameters._lower_limits[parameters._parind[:]],
+                    parameters._upper_limits[parameters._parind[:]])
             if outsidebounds is True:
                 next_set, outbound = set_outside_bounds(next_set=next_set)
                 trypath.append(next_set)
@@ -66,12 +67,8 @@ class DelayedRejection:
             # Evaluate new proposals
             outbound = 0
             # evaluate likelihood function
-            next_set.newlike = likeobj.evaluate_likelihood(
+            next_set.like = likeobj.evaluate_likelihood(
                     next_set.theta, custom=custom)
-            if isinstance(next_set.newlike, dict):
-                next_set.ss = next_set.newlike['ssq']
-                next_set.like = next_set.newlike['like']
-#            next_set.ss = sosobj.evaluate_sos_function(next_set.theta, custom=custom)
             # evaluate prior function
             next_set.prior = priorobj.evaluate_prior(theta=next_set.theta)
             trypath.append(next_set)  # add set to trypath
@@ -85,14 +82,13 @@ class DelayedRejection:
 
     # -------------------------------------------
     @classmethod
-    def initialize_next_metropolis_step(cls, npar, old_theta, sigma2, RDR):
+    def initialize_next_metropolis_step(cls, npar, old_theta, RDR):
         '''
[…]

[…]
