### 2019-05-08 8f59dee
pass in likelihood object which contains ssq
