+++ pymcmcstat/procedures/ErrorVarianceEstimator.py
@@ -37,7 +37,8 @@ class ErrorVarianceEstimator:
     def update(self, current_set, **kwargs):
         self.status = 'updating'
         self.counter += 1
-        sigma2 = self.update_error_variance(current_set.ss, self.model)
+        sigma2 = self.update_error_variance(
+                current_set.like['ssq'], self.model)
         self.sigma2[self.counter, :] = sigma2.reshape(sigma2.size, )
         return sigma2
 
