### 2019-05-08 10f79a5
added sigma2 to default likelihood object
