### 2019-05-08 55fbf35
removed old default times from set
- everything is inside of likelihood or prior dictionaries
