### 2019-05-13 9056c98
removed import statement of SOS class
