+++ pymcmcstat/__init__.py
@@ -1,1 +1,1 @@
-__version__ = "1.7.1"
+__version__ = "1.8.0rc1"
