### 2019-05-13 15050d9
commented out obsolete methods
