### 2019-05-13 6b62923
minor updates to MCMC
