### 2019-05-13 eb54ca3
updated default output from custom samplers to `None`
