### 2019-05-14 498beff
printed tests for debug
