### 2019-05-16 2860c1b
minor updates
- found the bug, not sure how to fix yet.  Sigma2 needs to be the same when comparing sum-of-squares.
