### 2019-05-18 d98effe
Updated link to new tutorials location
- Tutorials have been added to the pymcmcstat directory to facilitate outside contributions to the code.
