+++ README.rst
@@ -35,7 +35,7 @@ You can also clone the repository and run ``python  setup.py install``.
 Getting Started
 ===============
 
-- `Tutorial notebooks <https://nbviewer.jupyter.org/github/prmiles/notebooks/tree/master/pymcmcstat/index.ipynb>`_
+- `Tutorial notebooks <https://nbviewer.jupyter.org/github/prmiles/pymcmcstat/tree/master/tutorials/index.ipynb>`_
 - `Documentation <http://pymcmcstat.readthedocs.io/>`_
 - `Release history`_
 - `Contributing guidelines`_
