### 2019-05-18 9ce986c
Added v1.7.1 to changelog
