### 2019-05-30 33f3f00
Fixed #48
- User can specify "skip" kwarg for pairwise correlation.
