### 2019-05-30 f9c6a92
created method for determining subsample
- used skip, maxpoints, and nsimu to id subsample array
