### 2019-05-30 d12571b
implemented subsampling option in pairwise and chain panel
