### 2019-05-30 d45ac78
minor formatting adjustments
