### 2019-05-30 590d2f0
fixed formatting issues in plotting modules
