### 2019-05-30 70dae8f
added optional display for definition of items in chainstats
