### 2019-05-30 90c8ef3
updated changelog to reflect new feature of chainstats
