+++ CHANGELOG.rst
@@ -5,6 +5,7 @@ v1.8.0 ()
 ----------------------
 - Added acceptance rate display feature when calling chain statistics
 - User can specify `skip` or `maxpoints` in pairwise correlation and chain panel plots in order to thin chain.
+- User can request item definitions when callings `chainstats`.
 
 v1.7.0 (April 5, 2019)
 ----------------------
