### 2019-05-30 d52b1b0
added unit test for display of chainstats
- checks for display_details True and False
