### 2019-06-04 #1417 @katyhuff
Sorry for the delay here @prmiles. I've  to the prompt and thorough review by @djmitche Without the review  from @jarvist, I'll move forward with this review  on my own, as I have previously installed and checked the software myself. I'll review all the checks today and we can move forward.
