### 2019-06-04 #1417 @katyhuff
Thank you @djmitche for your review -- we couldn't do this without you.

Thank you @prmiles for a strong submission and for engaging actively in the review process! I have looked over the paper, double checked all the DOI links, and have conducted a high level review of the code itself. Everything looks ship-shape to me. At this point, please double check the paper yourself, review any lingering details in your code/readme/etc. that you intend to make, and then make an archive of the reviewed software in Zenodo/figshare/other service. Please be sure that the DOI metadata (title, authors, etc.) matches this JOSS submission. You may update the version to reflect these changes as necessary. If you do need to update the version please let me know the new version number. Once that's complete, please update this thread with the DOI of the archive, and I'll move forward with accepting the submission! Until then, now is your moment for final touchups!
