### 2019-06-04 eca6a7d
Fix mcmcstat link
- Fixed #56
