### 2019-06-05 #1417 @prmiles
The updated DOI is [10.5281/zenodo.2660145](http://doi.org/10.5281/zenodo.2660145).  This corresponds to version 1.7.1, which includes the corrections requested by @djmitche

[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.2660145.svg)](https://doi.org/10.5281/zenodo.2660145)

I fixed several items in the paper and I am completing my final touch-ups.  I'll let you know as soon as they are complete.  Thank you!
