### 2019-06-05 #1417 @katyhuff
@openjournals/joss-eics I believe this paper is ready for acceptance!
