### 2019-06-05 #1417 @prmiles
I have updated the Zenodo archive metadata title such that it matches the paper.

Regarding the ORCID, I have linked it to my Zenodo account.  Is that what you are requesting?  The ORCiD is in the markdown file for the paper as well.  I believe I formatted it correctly.

Thank you!
