### 2019-06-05 #1417 @labarba
Folks, I read the paper and I find it awfully hard ... I mean, it really is full of jargon and acronyms. Nobody commented on that during the review? JOSS does say in the Author guidelines that the article should be written aimed at a non-specialist technical audience. Do you want to give this a new look?
