### 2019-06-05 #1417 @prmiles
I apologize for the confusion regarding the intended audience.  I am happy to update the paper to communicate with a non-specialist technical audience, but I will need a few days to do this effectively.  Assuming the other reviewers are in agreement, I will begin updating the paper and hopefully have it ready by Monday.  Thank you for the feedback.
