### 2019-06-05 #1417 @labarba
The [relevant section](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain) of our documentation says:

> JOSS welcomes submissions from broadly diverse research areas. For this reason, we require that authors include in the paper some sentences that would explain the software functionality and domain of use to a non-specialist reader.

I really appreciate your willingness to give this another pass. It benefits JOSS and all our community if we try to communicate better about the software we're publishing. (And it's hard work to write about technical topics without jargon.)
