### 2019-06-05 #1417 @katyhuff
@labarba Thanks for this feedback. I did find it readable, but likely because I am quite familiar with this area of application. I'll be more careful about this jargon-blindness in the future.

@prmiles Thank you for being willing to make these changes. In addition to a reduction of acronyms, perhaps moving the 4th paragraph closer to the beginning of the paper would help to provide a softer onramp for the reader.
