### 2019-06-05 #1417 @prmiles
@katyhuff Thank you for the suggestion regarding the 4th paragraph!  I'm grateful the issue was brought up, and thank you for your patience as I make the updates.
