### 2019-06-05 2df7eef
Moved 4th paragraph to near begining
