### 2019-06-05 4f56b76
Added basic description of likelihood/prior
