### 2019-06-05 6db6f52
Updated posterior dist. eqn.
