+++ paper/paper.md
@@ -22,7 +22,7 @@ Many scientific problems require calibrating a set of model parameters in order 
 1. Parameters ($q$) are treated as random variables with underlying distributions instead of unknown but fixed values.
 2. Observations (data) are expected to be equal to the model response plus idependent and identically distributed random errors: $F^{data}(i) = F(i;q) + \epsilon_i$ where $\epsilon_i \sim \mathcal{N}(0,\sigma^2)$.
 
-The goal of the calibration process is now to infer the parameters' posterior distributions given a set of observations ($\pi(q|F^{data}(i))$).  Once these parameter distributions are known, the Bayesian approach provides a natural framework from which to consider how this uncertainty propagates and affects model prediction.
+The goal of the calibration process is now to infer the parameters' posterior distributions given a set of observations - $\pi(q|F^{data}(i))$.  Once these parameter distributions are known, the Bayesian approach provides a natural framework from which to consider how this uncertainty propagates and affects model prediction.
 
 The ``pymcmcstat`` package was designed for engineers and scientists interested in using Bayesian methods to quantify model parameter uncertainty.  Furthermore, we had the goal of providing a Python platform for researchers familiar with the MATLAB toolbox [mcmcstat](https://mjlaine.github.io/mcmcstat/) developed by Marko Laine.  To accommodate a diverse audience, we constructed several [tutorials](https://nbviewer.jupyter.org/github/prmiles/pymcmcstat/blob/master/tutorials/index.ipynb) to guide the user through the various stages of setting up a problem, such as defining the data structure, model parameters, and simulation options.  Currently, the package is limited to Gaussian likelihood and prior functions; however, these are still suitable for a wide variety of scientific problems.  As many users are not necessarily familiar with the statistical nomenclature behind the Bayesian approach, the package simply requires the user to define a function that calculates the sum-of-squares error with respect to the model and the observations (which is consistent with the second assumption listed above).  Knowledge know *a priori* about the parameter distributions is defined in the prior function; however, the default program behavior is to use uniform prior distributions which is a common approach for these types of problems.
 
