### 2019-06-05 7552d5c
Cleaned up sentences on prior
