### 2019-06-05 #1417 @djmitche
I'll look more closely at that in future reviews.  I did have to do a bit of background reading to understand the purpose of the package.
