+++ paper/paper.bib
@@ -59,10 +59,12 @@
 
 @article{zenodo,
   author= {Paul R. Miles},
-  title={prmiles/pymcmcstat: v1.7.1},
+  title={{pymcmcstat: A Python Package for Bayesian 
+                   Inference Using Delayed Rejection Adaptive
+                   Metropolis}},
   month=may,
   year=2019,
   doi={10.5281/zenodo.2660145},
   url={https://doi.org/10.5281/zenodo.2660145},
-  publisher={Zenodo}
+  publisher={Zenodo},
 }
