### 2019-06-06 #1417 @prmiles
I have modified the paper to accommodate a more general technical audience.  The first couple paragraphs provide a broad perspective of why the software is useful and where it can be used in scientific research (hopefully!).   My hope is that the first part of the paper gives the reader the big picture and enough information to know whether or not they could use the software.

The latter part of the paper still contains fairly technical content.  I added some discussion regarding Metropolis algorithms, and I believe the references provide a reasonable set of resources if someone wants to learn more about MCMC methods.

Ideally, the first part of the paper makes the technical content less overwhelming, but I am happy to continue iterating.  Thank you again for the feedback regarding the paper!
