### 2019-06-06 #1417 @labarba
Please delete the last sentence before the acknowledgements, and the citation to the Zenodo deposit. We ask that you don't include this as a citation, because the archive will be highlighted in the margin decorators on the first page of the paper (and the DOIs will be linked in the metadata of the paper).
