+++ paper/paper.bib
@@ -56,15 +56,3 @@
   url={https://doi.org/10.1198/jcgs.2009.06134},
   doi={10.1198/jcgs.2009.06134}
 }
-
-@article{zenodo,
-  author= {Paul R. Miles},
-  title={{pymcmcstat: A Python Package for Bayesian 
-                   Inference Using Delayed Rejection Adaptive
-                   Metropolis}},
-  month=may,
-  year=2019,
-  doi={10.5281/zenodo.2660145},
-  url={https://doi.org/10.5281/zenodo.2660145},
-  publisher={Zenodo},
-}
