### 2019-06-06 #1417 @prmiles
I have removed the final sentence, including the citation to the Zenodo deposit.  I also removed the Zenodo information from the .bib file.
