### 2019-06-06 #1417 @labarba
Hmm. Did you forget to push the change? I still see the sentence and citation in the PDF created with the `accept` command above.
