### 2019-06-06 #1417 @prmiles
The paper looks correct in the master version of my GitHub repository... I'm not entirely sure how the compiling process works with whedon.  Does whedon pull the latest version from the repository when you use the accept command?  Or only when you generate the pdf?
