### 2019-06-06 #1417 @whedon
:tada::tada::tada: Congratulations on your paper acceptance! :tada::tada::tada:
